rename interface and refactor useSingleQuotes
