+++ src/index.ts
@@ -20,6 +20,7 @@ export * from './makefile';
 export * from './node-package';
 export * from './node-project';
 export * from './object-file';
+export * from './option-hints';
 export * from './project';
 export * from './readme';
 export * from './sample-file';
