+++ src/javascript/index.ts
@@ -1,2 +1,1 @@
 export * from './projenrc';
-export * from './render-options';
