+++ src/option-hints.ts
@@ -1,0 +1,19 @@
+/**
+ * Choices for how to display commented out options.
+ */
+export enum OptionHints {
+  /**
+   * Display all possible options (grouped by which interface they belong to).
+   */
+  ALL = 'all',
+
+  /**
+   * Display only featured options, in alphabetical order.
+   */
+  FEATURED = 'featured',
+
+  /**
+   * Display no extra options.
+   */
+  NONE = 'none'
+}
