+++ src/typescript/projenrc.ts
@@ -1,7 +1,8 @@
 import { existsSync, writeFileSync } from 'fs';
 import { resolve } from 'path';
 import { Component } from '../component';
-import { OptionHints, renderJavaScriptOptions } from '../javascript/render-options';
+import { renderJavaScriptOptions } from '../javascript/render-options';
+import { OptionHints } from '../option-hints';
 import { TypeScriptProject } from '../typescript';
 
 export interface ProjenrcOptions {
@@ -13,7 +14,7 @@ export interface ProjenrcOptions {
 
   /**
    * Include commented out properties.
-   * @default ProjectOptionsVerbosity.FEATURED
+   * @default OptionHints.FEATURED
    */
   readonly comments?: OptionHints;
 }
